Update RequireVisitor.cs
When passing a parameter identifier instead of a string literal to a require call, the require visitor throws an exception.
While it is true that the visitor can not read the argument out of a static context, it can be resolved at runtime. 
So the visitor should ignore this case or maybe give a notification instead of throwing a general exception.